Add Timer component with component lifecycles
The Timer component has been created and has the sessionTime and
breakTime passed down as props, which it uses to create it's own state
for the timer. A lifecycle hook was also added to update the state
in this component when the timers change at the App component level.

The countdown timer will be handled within the state of the Timer
component.