Refactor countdown timer method and props updates
The countdown timer method was refactored to updated the millisecond
countdown state within the setInterval function so that it is always
in sync with the time left on the timer. Additionally the properties
updates for changing the sessionTime and breakTime were corrected as
they were causing the timers to restart on any property changes rather
than just monitoring changes within their own properties.