Update receive props rules for Timer.js
Unnecessary prop rules about changing the timers were removed as it is
now controlled by the currTimer state within the component. The old
code was causing some bugs and was more convoluted. Additionally, more
state manipulation was added for when the sessionTime and breakTime
are changed and passed down to the Timer component. Not only will it
change the countdown value in the Timer state, but will also update
the minutes and seconds state if the timer being manipulated is the
active timer.