Fix bug with timer initiating multiple times
The componentWillReceiveProps had a bug with regard to the starting
and pausing of the timer. It was only checking the nextProps it was
receiving and not comparing them to the current props. This caused
issues if the user tried to change the timer amount on one of the
counters while the timer was running. It would not change the time,
but would actualize more intervalIds and cause the countdown to not
be accurate and visually glitchy. The logic was updated to monitor
changing the boolean value which resolves the issue.