Refactor timer method and logic to switch timers
The logic to run the timers and alternate between the timers has been
greatly simplified. When the component updates it checks if one of the
timers has reached zero. If so, it changes the currTimer to the other
value and resets the timer countdowns for both timers.

Additionally the minutes and seconds were removed from state entirely.
This is now completed within the render method. This simplifies the
number of states that must be managed and also resolved a bug were
the seconds in state did not match the countdown timer. It also allowed
the removal of the date math code as the timer simply counts down the
ms stored in the current timer countdown state.